% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_journals.R
\name{with_journals}
\alias{with_journals}
\title{Evaluate an R expression while collecting journals from completed futures}
\usage{
with_journals(expr, substitute = TRUE, envir = parent.frame())
}
\arguments{
\item{expr}{The R expression to evaluate}

\item{substitute}{If TRUE, then \code{expr} is subtituted, otherwise not.}

\item{envir}{The environment where \code{expr} should be evaluated}
}
\value{
A list of FutureJournal:s.
}
\description{
Evaluate an R expression while collecting journals from completed futures
}
\examples{
library(future.apply)

slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(multisession, workers = 2)
js <- with_journals({
  vs <- future_lapply(3:1, FUN = slow_fcn)
})
print(js)

## Shut down parallel workers
plan(sequential)
}
