% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggjournal.R
\name{ggjournal}
\alias{ggjournal}
\title{Create a Future Journal Plot}
\usage{
ggjournal(futures, baseline = TRUE, ...)
}
\arguments{
\item{futures}{A list of \link[future:Future-class]{future::Future} objects.}

\item{baseline}{(POSIXct; optional) A timestamp to server as time zero
for the relative timestamps. If \code{TRUE} (default), then the earliest
timepoint observed is used as the baseline.}

\item{\ldots}{Currently not used.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Create a Future Journal Plot
}
\examples{
library(future)
oopts <- options(future.journal = TRUE)

slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(sequential)
fs <- lapply(5:1, FUN = function(x) future(slow_fcn(x)))
vs <- value(fs)
print(ggjournal(fs))

plan(multisession, workers = 2)
fs <- lapply(5:1, FUN = function(x) future(slow_fcn(x)))
vs <- value(fs)
print(ggjournal(fs))

## Decrease polling time to 0.01 seconds (default is 0.1 seconds)
options(future.wait.interval = 0.01)
fs <- lapply(5:1, FUN = function(x) future(slow_fcn(x)))
vs <- value(fs)
print(ggjournal(fs))

## Shut down parallel workers
plan(sequential)
options(oopts)
}
