% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggjournal.R
\name{ggjournal}
\alias{ggjournal}
\title{Create a Future Journal Plot}
\usage{
ggjournal(futures, baseline = NULL, ...)
}
\arguments{
\item{futures}{A list of \link[future:Future-class]{future::Future} objects.}

\item{baseline}{(optional) A POSIXct timestamp to server as time zero
for the relative timestamps.}

\item{\ldots}{Currently not used.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Create a Future Journal Plot
}
\examples{
library(future)
plan(sequential)

t_start <- Sys.time()
fs <- lapply(5:1, FUN = function(x) future({
  Sys.sleep(x / 10)
  sqrt(x)
}))
vs <- value(fs)

gg <- ggjournal(fs, baseline = t_start)
print(gg)
}
